Gemini-GCP/requirements.txt
google-generativeai==0.8.3
pillow==11.0.0